🐛 fix google workspace discovery (#907)
This PR fixes the platform detection for google workspace, a bug
introduced with https://github.com/mondoohq/cnquery/pull/879

Now it returns the right things for gcp and google workspace

```javascript
$ cnquery shell googleworkspace --customer-id <customerid> --impersonated-user-email <email>
cnquery> asset { name family title runtime platform }
asset: {
  title: "Google Workspace"
  platform: "google-workspace"
  runtime: "google-workspace"
  name: "Google Workspace mondoo.com"
  family: [
    0: "google"
  ]
}
```


```
$ cnquery shell gcp
  ___ _ __   __ _ _   _  ___ _ __ _   _ 
 / __| '_ \ / _` | | | |/ _ \ '__| | | |
| (__| | | | (_| | |_| |  __/ |  | |_| |
 \___|_| |_|\__, |\__,_|\___|_|   \__, |
  mondoo™      |_|                |___/  interactive shell

cnquery> asset { name family title runtime platform }
→ platform override: gcp-project
asset: {
  title: "GCP Project"
  platform: "gcp-project"
  runtime: "gcp"
  name: "GCP project example-test-12345"
  family: [
    0: "google"
  ]
}
```
